@@ -2,8 +2,8 @@ package org.zamahaka.cheremosh.extensions
 
 import android.content.Context
 import android.graphics.drawable.Drawable
-import android.support.annotation.*
-import android.support.v7.widget.RecyclerView
+import androidx.annotation.*
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
