@@ -1,7 +1,7 @@
 package org.zamahaka.cheremosh.ui
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_main.*
 import org.zamahaka.cheremosh.R
 import org.zamahaka.cheremosh.ui.concert.ConcertListFragment
