@@ -1,8 +1,8 @@
 package org.zamahaka.cheremosh.ui.rv
 
-import android.support.v7.recyclerview.extensions.ListAdapter
-import android.support.v7.util.DiffUtil
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
 import org.zamahaka.cheremosh.model.IdEntity
 
 abstract class IdEntityRvAdapter<T : IdEntity<*>, VH : RecyclerView.ViewHolder>(
