@@ -1,7 +1,7 @@
 package org.zamahaka.cheremosh.ui.notes
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import android.util.Log
 import com.google.firebase.database.*
 import org.zamahaka.cheremosh.model.NoteFile
